ChatGPT4 generated code for a Billiard simulation
Prompt : "please write a web page using babylon.js and ammo.js as the physics engine that creates a solid ground of color green and dimensions 1.27 m by 2.54 m, creates walls of color green and height 0.05 m on all the edges of the ground and parallel to the corresponding edge, creates three spheres of white color, diameter 0.062 m and mass 0.22 Kg. When the user clicks on the surface of one of the spheres, the clicked-on sphere will increase its velocity in 1.0 m/s in the direction parallel to a line going from the point of view of the camera to the point of intersection with the clicked sphere. The collisions between spheres have a restitution coefficient of 0.98. The collisions. between spheres and walls have a restitution coefficient of 0.98. The collisions between the spheres and the ground plane have a restitution coefficient of 0.8. The rolling friction coefficient between sphere and ground is 0.06. Add shadows to the scene, including shadows projected from the spheres to the ground plane. "

Note, results may be totally different for the same prompt as ChatGPT is not deterministic